# fix(anta): Add Semaphore to AsyncEOSDevice