@gmuloc 2025-02-21
that being said - would be cool if we manage to get a benchmark for this kind of stuff running somewhere.. :)